Add authorization to create, edit and delete property

+++ app/controllers/admin/properties_controller.rb
@@ -1,8 +1,12 @@
+require_relative "../../policies/property_policy"
+
 class Admin::PropertiesController < ApplicationController
+  include Pundit::Authorization
+
   layout "admin"
 
   expose :property
-  expose :properties, -> { Property.all }
+  expose :properties, -> { current_user.properties }
 
   before_action :authenticate_user!
 
@@ -10,20 +14,37 @@ class Admin::PropertiesController < ApplicationController
     if property.save
       redirect_to admin_properties_path
     else
-      flash[
-        :error
-      ] = "Error creating property: #{property.errors.full_messages.to_sentence}"
+      flash[:error] = "Error creating property: #{property.errors.full_messages.to_sentence}"
       redirect_to new_admin_property_path
     end
   end
 
+  def show
+    authorize property
+  end
+
+  def edit
+    authorize property
+  end
+
+  def update
+    authorize property
+
+    if property.update(property_params)
+      redirect_to admin_properties_path
+    else
+      flash[:error] = "Error updating property: #{property.errors.full_messages.to_sentence}"
+      redirect_to edit_admin_property_path(property)
+    end
+  end
+
   def destroy
+    authorize property
+
     if property.destroy
       flash[:success] = "Property deleted"
     else
-      flash[
-        :error
-      ] = "Error deleting property: #{property.errors.full_messages.to_sentence}"
+      flash[:error] = "Error deleting property: #{property.errors.full_messages.to_sentence}"
     end
 
     redirect_to admin_properties_path
@@ -34,6 +55,7 @@ class Admin::PropertiesController < ApplicationController
   def property_params
     params.require(:property).permit(
       :id,
+      :user_id,
       :property_type,
       :price_cents,
       :area,
